Migrate to vscode.
