+++ pubspec.yaml
@@ -33,7 +33,7 @@ dependencies:
   google_fonts: ^2.1.0
   flutter_svg: ^0.23.0+1
   flutter_gen: ^4.1.3
-
+  
   window_size:
     git:
       url: git://github.com/google/flutter-desktop-embedding.git
